pyproject.toml
[project]
name = "course-building-aws-agent-core"
version = "0.1.0"
description = "AWS Bedrock AgentCore sample using Strands agents, managed with uv."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "bedrock-agentcore==0.1.1",
  "strands-agents==1.4.0",
  "strands-agents-tools==0.2.3",
  "boto3>=1.34",
  "requests>=2.31",
  "mcp>=1.12.0",
  "numpy>=2.2.6",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]

[dependency-groups]
dev = [
  "bedrock-agentcore-starter-toolkit==0.1.5",
  "black>=25.1.0",
  "flake8>=7.3.0",
  "isort>=6.0.1",
]


[tool.flake8]
max-line-length = 100

